+++ client/src/CountriesList.js
@@ -1,0 +1,8 @@
+import React from 'react'
+
+const CountriesList = () => {
+    return (
+        <h1>Countries list!</h1>
+    )
+}
+export default CountriesList
